Fix directory structure, missing URLs and files
