+++ react-app/src/pages/NoPage.tsx
@@ -1,0 +1,10 @@
+const NoPage = () => {
+    return (
+        <div>
+            <h1>404</h1>
+            <p>Page Not Found</p>
+        </div>
+    );
+};
+
+export default NoPage;
